fix(ci): refactor FactoryGenerator for macOS 10.13 and resolve all warnings
This commit addresses the macOS build failure and eliminates all remaining compiler warnings:
- Refactored FactoryGenerator to use juce::File instead of std::filesystem for compatibility with macOS 10.13.
- Fixed syntax for [[maybe_unused]] attributes.
- Resolved variable shadowing in PluginProcessor, PluginEditor, and UI sections.
- Fixed member initialization order in PluginEditor.
- Added missing override keywords.
- Resolved implicit sign and float conversion warnings.

+++ src/ui/sections/ControlSection.cpp
@@ -52,7 +52,7 @@ ControlSection::ControlSection()
 
             // Map 0..100 to 0..5
             int index = juce::jlimit(0, 5, (int)(value / 16.66f));
-            return rateDivisions[index];
+            return rateDivisions[static_cast<size_t>(index)];
         };
 
     rateKnob.onValueChanged = [this](double value) {
